pyproject.toml
[project]
name = "vesta-rental-index"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2,<5.3",
    "django-ninja>=1.3,<2.0",
    "psycopg[binary]>=3.2,<4.0",
    "gunicorn>=23.0,<24.0",
    "requests>=2.32,<3.0",
]
